refactor: 貸出シートを子Container(BorrowSheetContainerView)に分割
BorrowListContainerViewに同居していた図書一覧画面と貸出シートの状態を
「画面の境界＝状態の境界」に沿って分離する。シート内フロー
(だれが借りますか？→どの枠で借りますか？→✓カード)の一切を
BorrowSheetContainerViewへ移し、シート表示ごとの@Stateの生成・破棄に
初期化を任せることで、openBorrowSheetでの手動リセット
(sheetPath/selectedClassGroupId)を廃止する。

- 子はEnvironmentで必要なModelを自分で取得し、親→子はcontext引数と
  onClose/onLendCompletedクロージャのみ
- BorrowSheetContextのスナップショット焼き込み(#186修正)と理由コメントは維持
- 3タップ貸出フロー・フォーム/ページシート分岐・✓カード2.5秒後の自動クローズ・
  無操作15秒の置き去り復帰・枠入れ替えUndo・誤スワイプ防止・
  貸出完了時の絞り込みリセット/先頭スクロールはすべて挙動不変
